Add initial version of the Haskell Catalog
This commit adds the HaskellCatalog class, which loads sets of available
functions (divided into categories) from a bundled XML file. A working
XML file has been bundled, as well as an XML Schema definition file that
can be used (and is used by the HaskellCatalog class) to check if the
XML data file is valid.

Functions can either be requested by name or by category, and are
returned as Entry objects which can have a short documentation comment.