Update TactileAPI which is now called TactileFX
The TactileAPI library underwent a slight update and is now called
TactileFX.

The [upstream][] repository also now contains pom.xml files for the
components, which means we could theoretically use TactileFX as a Maven
dependency. However, the pom.xml has not yet been published to Maven
Central, and using it as a dep would make modification and
experimentation harder. To make a long story short, this is just a plain
copy for now.

[upstream]: https://github.com/Frumbasjt/TactileAPI