Implement panning (drag) and zooming (scroll)
This commit adds panning and zooming functionality.

Pan is implemented by translating the TactilePane in the correct (that
is, opposite) direction. Similarly, zooming is implemented by scaling
the TactilePane.

The code for TactilePane is changed to use pane-relative instead of
scene-relative coordinates at some points, as scene-relative coordinates
do not take into account the transformations applied.

Touch interaction is untested.