Have ComponentLoader swallow/rethrow IOExceptions
This changes the ComponentLoader interface to itself catch IOExceptions
and rethrow them as RuntimeExceptions.

Not throwing IOExceptions simplifies quite a few methods, as well as the
MainMenu class.

Rationale: When an FXML file can not be read, that's *always* a fatal
error. In other words, the program is useless when it does not have all
its FXML files. In the old implementation, none of the places that
caught the exception did something useful with it.