Make GhciSession a Service that returns Futures
This enables GhciSession to execute Haskell snippets on a different
thread than the main thread.

GhciSession is an AbstractExecutionThreadService from Guava. It takes Haskell
source to execute off the work queue, which is read from a (singular)
execution thread, and filled by anything that wants something executed.

Push and pull methods on GhciSession return 'immediately', and never
throw HaskellException. If (when) something goes wrong, the future will
be set to unsuccesful, which means `get` will throw ExecutionException
or the unsuccesful callback will be called.

DisplayBlock uses callbacks to delay updating of its label until the
future is completed (i.e. evaluation should not block the UI).

RGBBlock and GraphBlock wait for their results for now (i.e. their
behaviour should be unchanged and they will block the UI).