Use a TopMenu on Mac OS X instead of onscreen context menu for touch devices.
Added MacTopMenu to create a Top application menu for Mac.
GlobalMenu refactored into GlobalContextMenu that is used for on-screen context menu, and MenuActions which are shared between all menus.
MainOverlay now wraps overlay in a BorderPane that accepts a TopMenu, and when detected we are running on a Mac it replaces the on-screen
context menu with a TopMenu that does the same actions (except full-screen which is a window control in Mac OS X).

Avoid referencing overlay via Main.java public overflow, use the parameter passed in to the constructor and pass down to refreshTheme()
MainOverlay doesn't need to be public in Main.java any more.