Zoom actions as View menu items with shortcuts.
Made ZoomIn and ZoomOut MenuActions.
Added menu items in View menu with shortcuts for ZoomIn and ZoomOut.
Moved the zoom() method to ToplevelPane which is in fact where it is applied.
Moved all actions into MenuActions and decoupled from MainOverlay as in fact it's not needed for many actions.

NOTE: For some strange Java/JavaFX reason, the Keycode.MINUS short-cut doesn't work, at least on Mac.