Second attempt at serialization and deserialization. See below for issues I could not get working.
New
===
Now attempts to create a CatalogFunction from the catalog using the name only.
Better use of constants for serialization keys.
Added serialization of the format version number, loading it and checking compatible on deserialization.
Sets the window title to be the filename when saved/saved-as or "New file" when new.

Problems:
Creating Types deserializing constant blocks. See ConstantBlock.java for this comment:
        // FIXME Use Environment/TypeBuilder to create Type from String?
        // FIXME I got lost in the code trying to figure out how to do this isung DataTypeInfo, TypeCon, etc...
        // FIXME Type type = (String)bundleFragment.get("type");
 Easy to reproduce the problem opening the samples/simple/hello_world.viskell sample