Moved two methods that perform Menu actions that were very self-contained from MainOverlay.java into MenuActions.java.
Moved the zoom() method into ToplevelPane() which is the Pane it actually changes the view of and hence belongs to.

Added Keyboard shortcuts for most menu items.
NOTE: There is a strange bug on Mac that Command & '-' to zoom out doesn't work on JavaFX.

Added unit tests for some of the menu actions